dev-requirements.txt
docker
discord.py
mcrcon
